Megatron-FSDP: extract allocator classes into allocators.py
Move Bucket, TemporaryBucketAllocator, StorageResizeBasedBucketAllocator,
RotaryBucketAllocator, FixedPoolAllocator, and MultiGroupUBRAllocator (plus
the private _p_assert/_alloc_storage/_free_storage helpers) out of
param_and_grad_buffer.py into a new allocators.py, and re-import them from
param_and_grad_buffer.py to preserve the existing public surface. The
nccl_allocator shim stays in param_and_grad_buffer.py untouched; to avoid a
module-level import cycle, MultiGroupUBRAllocator.__init__ imports it lazily,
and FixedPoolAllocator keeps its ParameterGroup annotation under
TYPE_CHECKING.

No behavior change — pure code move.
